Refactor database rule config refresh logic
