+++ features/readwrite-splitting/core/src/main/java/org/apache/shardingsphere/readwritesplitting/event/ReadwriteSplittingRuleConfigurationEventBuilder.java
@@ -18,19 +18,10 @@
 package org.apache.shardingsphere.readwritesplitting.event;
 
 import com.google.common.base.Strings;
-import org.apache.shardingsphere.infra.config.algorithm.AlgorithmConfiguration;
+
 import org.apache.shardingsphere.infra.rule.event.GovernanceEvent;
-import org.apache.shardingsphere.infra.util.yaml.YamlEngine;
 import org.apache.shardingsphere.mode.event.DataChangedEvent;
-import org.apache.shardingsphere.mode.event.DataChangedEvent.Type;
 import org.apache.shardingsphere.mode.event.config.RuleConfigurationEventBuilder;
-import org.apache.shardingsphere.readwritesplitting.api.rule.ReadwriteSplittingDataSourceRuleConfiguration;
-import org.apache.shardingsphere.readwritesplitting.event.config.AddReadwriteSplittingConfigurationEvent;
-import org.apache.shardingsphere.readwritesplitting.event.config.AlterReadwriteSplittingConfigurationEvent;
-import org.apache.shardingsphere.readwritesplitting.event.config.DeleteReadwriteSplittingConfigurationEvent;
-import org.apache.shardingsphere.readwritesplitting.event.loadbalance.AddLoadBalanceEvent;
-import org.apache.shardingsphere.readwritesplitting.event.loadbalance.AlterLoadBalanceEvent;
-import org.apache.shardingsphere.readwritesplitting.event.loadbalance.DeleteLoadBalanceEvent;
 import org.apache.shardingsphere.readwritesplitting.metadata.converter.ReadwriteSplittingNodeConverter;
 
 import java.util.Optional;
@@ -45,38 +36,6 @@ public final class ReadwriteSplittingRuleConfigurationEventBuilder implements Ru
         if (!ReadwriteSplittingNodeConverter.isReadwriteSplittingPath(event.getKey()) || Strings.isNullOrEmpty(event.getValue())) {
             return Optional.empty();
         }
-        Optional<String> groupName = ReadwriteSplittingNodeConverter.getGroupName(event.getKey());
-        if (groupName.isPresent() && !Strings.isNullOrEmpty(event.getValue())) {
-            return createReadwriteSplittingConfigEvent(databaseName, groupName.get(), event);
-        }
-        Optional<String> loadBalancerName = ReadwriteSplittingNodeConverter.getLoadBalancerName(event.getKey());
-        if (loadBalancerName.isPresent() && !Strings.isNullOrEmpty(event.getValue())) {
-            return createLoadBalanceEvent(databaseName, loadBalancerName.get(), event);
-        }
-        return Optional.empty();
-    }
-    
-    private Optional<GovernanceEvent> createReadwriteSplittingConfigEvent(final String databaseName, final String groupName, final DataChangedEvent event) {
-        if (Type.ADDED == event.getType()) {
-            return Optional.of(new AddReadwriteSplittingConfigurationEvent<>(databaseName,
-                    YamlEngine.unmarshal(event.getValue(), ReadwriteSplittingDataSourceRuleConfiguration.class)));
-        }
-        if (Type.UPDATED == event.getType()) {
-            return Optional.of(new AlterReadwriteSplittingConfigurationEvent<>(databaseName, groupName,
-                    YamlEngine.unmarshal(event.getValue(), ReadwriteSplittingDataSourceRuleConfiguration.class)));
-        }
-        return Optional.of(new DeleteReadwriteSplittingConfigurationEvent(databaseName, groupName));
-    }
-    
-    private Optional<GovernanceEvent> createLoadBalanceEvent(final String databaseName, final String loadBalanceName, final DataChangedEvent event) {
-        if (Type.ADDED == event.getType()) {
-            return Optional.of(new AddLoadBalanceEvent<>(databaseName, loadBalanceName,
-                    YamlEngine.unmarshal(event.getValue(), AlgorithmConfiguration.class)));
-        }
-        if (Type.UPDATED == event.getType()) {
-            return Optional.of(new AlterLoadBalanceEvent<>(databaseName, loadBalanceName,
-                    YamlEngine.unmarshal(event.getValue(), AlgorithmConfiguration.class)));
-        }
-        return Optional.of(new DeleteLoadBalanceEvent(databaseName, loadBalanceName));
+        return Optional.of(new ReadwriteSplittingRuleConfigurationChangedEvent(databaseName));
     }
 }
